backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1  # <-- ES VITAL QUE SEA ESTA VERSIÓN
python-multipart==0.0.9
email-validator==2.2.0
apscheduler==3.10.4
pypdf[crypto]==6.6.2  # Para manejo completo de PDFs (lectura, desencriptación, escritura)
reportlab==4.2.5  # Generación de PDFs de propuestas
httpx==0.27.0  # Cliente HTTP asíncrono para APIs externas (BanRep, Socrata)
google-genai>=1.0.0  # Google Gemini AI SDK moderno con Client()